64-bit fixes (#686)
* wip

* VIRTUAL_TO_PHYSICAL2

* engine

* ptr macro

* level commands C

* level_commands asm

* Behavior scripts

* unused object fields

* Animation clean-up

* Portable animations

* remove unnecessary void* casts

* engine

* goddard, libultra

* More fixes

* .balign 4 in glabel macro

* mbi

* geo fixes

* animations

* .dword for bin IDs

* better naming

* .word's

* .word32

* remove .ptr

* .long -> .word

* type fix

* Fix JP

* Remove unnecessary casts

* Remove some hard-coded sizes

* Animation fix

* Avoid hard-coded sizes in memory.c

+++ src/audio/memory.c
@@ -109,7 +109,7 @@ void *soundAlloc(struct SoundAllocPool *pool, u32 size) {
 }
 
 void func_80316094(struct SoundAllocPool *pool, void *arg1, u32 arg2) {
-    pool->cur = pool->start = (u8 *) (((u32) arg1 + 0xf) & -0x10);
+    pool->cur = pool->start = (u8 *) (((uintptr_t) arg1 + 0xf) & -0x10);
     pool->size = arg2;
     pool->unused = 0;
 }
